+++ src/main/kotlin/com/atvgstudios/klib/extensions/Integer.kt
@@ -1,0 +1,3 @@
+package com.atvgstudios.klib.extensions
+
+fun Int.timesAs(char: String): String = char.repeat(this)