+++ src/main/kotlin/com/atvgstudios/klib/functions/TimeMillis.kt
@@ -1,0 +1,8 @@
+package com.atvgstudios.klib.functions
+
+/**
+ * Get the current system time in milli seconds
+ * @see System.currentTimeMillis
+ * @return
+ */
+fun currentMillis(): Long = System.currentTimeMillis()