+++ src/main/kotlin/com/atvgstudios/klib/kLibInf.kt
@@ -1,12 +1,9 @@
 package com.atvgstudios.klib
 
-import com.atvgstudios.klib.annotations.Author
-
-@Author(name = "Thomas Obernosterer")
 object kLibInf {
     const val name = "ATVG-Studios Kotlin Std Library"
-    const val versionId = 2 // A unique version number used to identify releases and to make easy requirement checks
-    const val version = "0.1.1"
+    const val versionId = 3 // A unique version number used to identify releases and to make easy requirement checks
+    const val version = "0.1.2"
     const val company = "ATVG-Studios"
     const val license = "MIT"
     val authors = arrayListOf(