+++ CHANGELOG.md
@@ -1,13 +1,42 @@
-# CHANGELOG
+# Changelog
+All notable changes to this project will be documented in this file.
 
-## Release Tree 0.x
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### Releases of 0.1.x
+## Unreleased
 
-#### Release 0.1.1 (21st April 2019)
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
 
-* Lint: Fix formatting
+## 0.1.2 - 23.04.2019
+### Added
+- Integer `timesAs` Extension
+- Top-Level `currentMillis` Function
+- `Hasher` Interface
+- `Base58` Object
+- `Sha256` Object
+- `Json`   Object
+- Type Class `HashResult`
 
-#### Release 0.1.0 (21st April 2019)
+### Changed
+- Moved `UniqueID` Object from Types to Objects package
 
-* This is the initial release. A superset of features is already incuded like the Word datatype, Byte helper extensions and more.
+## 0.1.1 - 21.04.2019
+### Fixed
+- Sematic issues shown by lint
+
+## 0.1.0 - 21.04.2019
+### Added
+- First implementation
+
+## Overview
+
+* Unreleased - [Diff 0.1.2 to master](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/0.1.1...master)
+* 0.1.2 - [Diff 0.1.1 to 0.1.2](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/0.1.1...0.1.2)
+* 0.1.1 - [Diff 0.1.0 to 0.1.1](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/0.1.0...0.1.1)
+* 0.1.0 - [0.1.0](https://gitlab.atvg-studios.at/atvg-studios/kLib/tags/0.1.0)